Added 'splashkill' program.
This simple program is meant to be launched directly by the kernel,
before min-init. It forks itself and sleeps until the ENTER key is
pressed ; when it occurs, it enables the framebuffer console by
writting to a file in /sys. This permits to 'kill' the kernel splash
screen, in order to display the boot log.